Move tests back into upper directory
