+++ pokemonbattlelib_suite_test.go
