+++ rng_test.go
